+++ CHANGELOG.md
@@ -16,8 +16,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Perf: Avoid parsing/joining URIs when not necessary, e.g., for relations without variables or absolute URIs.
 - Perf: Reuse libcurl handles in the Ethon adapter
 - Perf: Optimize libcurl option preparation in the Ethon adapter
+- Perf: Optimize parsing of data and relations
 
 ### Fixes
+
+- Only create relations from `*_url` fields with string values, e.g. not from numbers, objects, or arrays
 
 ### Breaks